Restructure HTTP integration
